+++ src/frontend/org/voltdb/TupleStreamStateInfo.java
@@ -21,14 +21,14 @@ public class TupleStreamStateInfo {
     public final DRLogSegmentId partitionInfo;
     public final boolean containsReplicatedStreamInfo;
     public final DRLogSegmentId replicatedInfo;
-    public final byte drVersion;
+    public final int drVersion;
 
-    public TupleStreamStateInfo(DRLogSegmentId partitionInfo, byte drVersion) {
+    public TupleStreamStateInfo(DRLogSegmentId partitionInfo, int drVersion) {
         this(partitionInfo, new DRLogSegmentId(Long.MIN_VALUE, Long.MIN_VALUE, Long.MIN_VALUE), drVersion);
     }
 
     public TupleStreamStateInfo(DRLogSegmentId partitionInfo,
-            DRLogSegmentId replicatedInfo, byte drVersion) {
+            DRLogSegmentId replicatedInfo, int drVersion) {
         this.partitionInfo = partitionInfo;
         this.replicatedInfo = replicatedInfo;
         this.containsReplicatedStreamInfo = (this.replicatedInfo.drId > Long.MIN_VALUE);
